contests-page: in progress
